+++ lib/money.rb
@@ -4,3 +4,4 @@ require "i18n" rescue LoadError
 require "money/currency"
 require "money/money"
 require "money/core_extensions"
+require "money/deprecations"