+++ spec/bank/variable_exchange_spec.rb
@@ -145,7 +145,7 @@ describe Money::Bank::VariableExchange do
 
     context "with :file provided" do
       it "writes rates to file" do
-        f = mock('IO')
+        f = double('IO')
         File.should_receive(:open).with('null', 'w').and_yield(f)
         f.should_receive(:write).with(JSON.dump(@rates))